used RPE enum

+++ src/main/java/com/solvd/automation_homework/IWorkout.java
@@ -1,5 +1,5 @@
 package com.solvd.automation_homework;
 
 public interface IWorkout{
-    Session completeSession(Workout workout,int rpe) throws InvalidIntensityException;
+    Session completeSession(Workout workout,double rpe) throws InvalidIntensityException;
 }
